setup new tool getErrorsTool. This tool takes 5 arguments. A service, a lmit, a lookback, an operation and a min_duration. When no service is specified this tool will use list_services to find all given services to offer those to be used as a service. if one service is chosen then that service will implement the jaeger v3/api for specific query parameters. These query params can have key value pairs for the following 1.query.service_name 2. query.start_time_min 3. query.start_time_max 4. query.attributes_error 5. query,search_depth 6. query.operation_name 7. query.duration_min
These queries will be used to get relevant data about the
particular service.

if all services are selected then a similar process will occur but this
will be done for each service in the application.
when all services are checked for errors there will be a summary which includes the following
1. total (numnber of services)
2. with_errors (number of services with errors)

There will also be a errors_by_service which will give the amount of errors per service.

Also added a types folder

Co-authored-by: bwalkerq <benjamin.q.walker@gmail.com>

# Conflicts:
#	retriever_mcp/src/server.ts
#	retriever_mcp/src/tools/index.ts